fix eq/ord/hash issues
Eq/Ord/Hash traits were wrongly implemented with assumption that
`a.as_str() == b.as_str()` => `a == b`, but this assumption is wrong, because
it doesn't take the indices into account. The new impl does.

This also means that we can't implement `Borrow<str>` because of it's strict
requirements on Eq/Ord/Hash traits. `AsRef<str>` is now implemented instead.